+++ src/fixe/convert/fixe64_fromfloat.c
@@ -4,12 +4,12 @@
 
 #include "../../fixe.h"
 
-fixe64_t fixe64_fromfloat(float a, uint8_t Q)
+fixe64_t fixe64_fromfloat(double a, uint8_t Q)
 {
     fixe64_t b;
 
     b.Q=(Q>58)?58:Q;
-    b.val=(int32_t)round(a * (float)(1<<b.Q));
+    b.val=(int64_t)round(a * (double)((int16_t)1<<b.Q));
 
     return b;
 }
